# specs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API enables end-users, owners and Sally devices to interact  with the central database.",
    "version": "1.0",
    "title": "Pantry REST API",
    "termsOfService": "http://casabots.com/terms/",
    "contact": {},
    "license": {
      "name": "Copyright (c) CASABOTS 2016"
    }
  },
  "host": "localhost:8080",
  "basePath": "/pantry",
...
  "paths": {
...
    "/portions": {
      "post": {
        "tags": [
          "portion-controller"
        ],
        "summary": "createPortion",
        "operationId": "createPortionUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "portion",
            "description": "portion",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Portion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Portion"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/portions/update": {
      "post": {
        "tags": [
          "portion-controller"
        ],
        "summary": "updatePortion",
        "operationId": "updatePortionUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "portion",
            "description": "portion",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Portion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Portion"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/portions/{portionId}": {
      "get": {
        "tags": [
          "portion-controller"
        ],
        "summary": "getPortion",
        "operationId": "getPortionUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "portionId",
            "in": "path",
            "description": "portionId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Portion"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "portion-controller"
        ],
        "summary": "deletePortion",
        "operationId": "deletePortionUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "portionId",
            "in": "path",
            "description": "portionId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/salads": {
      "post": {
        "tags": [
          "salad-controller"
        ],
        "summary": "createSalad",
        "operationId": "createSaladUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "salad",
            "description": "salad",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Salad"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Salad"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/salads/custom": {
      "post": {
        "tags": [
          "salad-controller"
        ],
        "summary": "createCustomSalad",
        "operationId": "createCustomSaladUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "salad",
            "description": "salad",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Salad"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Salad"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/salads/update": {
      "post": {
        "tags": [
          "salad-controller"
        ],
        "summary": "updateSalad",
        "operationId": "updateSaladUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "salad",
            "description": "salad",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Salad"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Salad"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/salads/{saladId}": {
      "get": {
        "tags": [
          "salad-controller"
        ],
        "summary": "getSalad",
        "operationId": "getSaladUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "saladId",
            "in": "path",
            "description": "saladId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Salad"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "salad-controller"
        ],
        "summary": "deleteSalad",
        "operationId": "deleteSaladUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "saladId",
            "in": "path",
            "description": "saladId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Portion": {
      "type": "object",
      "properties": {
        "calories": {
          "type": "integer",
          "format": "int32"
        },
        "id": {
          "type": "string"
        },
        "ingredientId": {
          "type": "string"
        },
        "sizeId": {
          "type": "string"
        }
      }
    },
    "Salad": {
      "type": "object",
      "properties": {
        "available": {
          "type": "boolean"
        },
        "createdBy": {
          "type": "string"
        },
        "customSalad": {
          "type": "boolean"
        },
        "imageUrl": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "portions": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "saladId": {
          "type": "string"
        }
      }
    }
  }
}